+++ telicent-graph-schema/src/main/java/io/telicent/jena/graphql/fetchers/telicent/graph/StateRelationshipsFetcher.java
@@ -12,16 +12,13 @@
  */
 package io.telicent.jena.graphql.fetchers.telicent.graph;
 
-import graphql.com.google.common.collect.Streams;
 import graphql.schema.DataFetcher;
 import graphql.schema.DataFetchingEnvironment;
 import io.telicent.jena.graphql.schemas.telicent.graph.models.TelicentGraphNode;
 import io.telicent.jena.graphql.schemas.telicent.graph.models.NonDirectionalRelationship;
 import io.telicent.jena.graphql.schemas.telicent.graph.models.State;
-import org.apache.jena.graph.Node;
 import org.apache.jena.sparql.core.DatasetGraph;
 import org.apache.jena.sparql.core.Quad;
-import org.apache.jena.vocabulary.RDF;
 
 import java.util.List;
 import java.util.stream.Collectors;
@@ -31,18 +28,13 @@ import java.util.stream.Stream;
  * A GraphQL {@link DataFetcher} that finds the relationships for a state
  */
 public class StateRelationshipsFetcher
-        extends AbstractLimitOffsetPagingFetcher<State, Quad, List<NonDirectionalRelationship>> {
+        extends AbstractStateRelationshipsFetcher<List<NonDirectionalRelationship>> {
 
     /**
      * Creates a new fetcher that finds the relationships involving states
      */
     public StateRelationshipsFetcher() {
 
-    }
-
-    @Override
-    protected Stream<Quad> select(DataFetchingEnvironment environment, DatasetGraph dsg, State state) {
-        return Streams.concat(outbound(dsg, state), inbound(dsg, state));
     }
 
     @Override
@@ -53,20 +45,6 @@ public class StateRelationshipsFetcher
                                                                      source.getStateNode().equals(q.getSubject()) ?
                                                                      q.getObject() : q.getSubject(), dsg.prefixes())))
                     .collect(Collectors.toList());
-    }
-
-    private static Stream<Quad> outbound(DatasetGraph dsg, State target) {
-        return dsg.stream(Node.ANY, target.getStateNode(), Node.ANY, Node.ANY)
-                  .filter(q -> q.getObject().isURI() || q.getObject().isBlank())
-                  .filter(q -> !q.getPredicate().equals(RDF.type.asNode()) && !q.getObject()
-                                                                                .equals(target.getEntityNode()));
-    }
-
-    private static Stream<Quad> inbound(DatasetGraph dsg, State target) {
-        return dsg.stream(Node.ANY, Node.ANY, Node.ANY, target.getStateNode())
-                  .filter(q -> q.getSubject().isURI() || q.getSubject().isBlank())
-                  .filter(q -> dsg.contains(Node.ANY, q.getSubject(), RDF.type.asNode(), Node.ANY));
-
     }
 
 }