+++ CHANGELOG.md
@@ -3,7 +3,7 @@
 # 0.10.0
 
 - Telicent Graph Schema improvements:
-    - Added new `relCounts` property to `Node` type to allow summarising available relationships
+    - Added new `relCounts` property to `Node` and `State` types to allow summarising available relationships
     - Added `limit` and `offset` parameters to various list fields, in combination with the new `relCounts` field this
       allows paging through the results